refactor(ripgrep): inline schemas into index.ts
